Update for Photo Gallery Collection View
Photos from the Photo Gallery UICollection are now able to be selected,
deselected and viewed. When selected, they will have a green border.
Included option to deselect all photos at once. Gallery photos now load
asynchronously; should not be slowing the device down or taking a while
to load. Implemented iCloud library functionality -- if photo is not
locally stored on the device, but rather on iCloud, application will
wait for the photo to load and display a loading progress bar.