Constants and Login Page
Created a Constants class. This class will be used for keep track of
key strings and other constant values throughout the application.

Login now checks to see if this is the first time the app has been
used. If so, it sets up some of the internal folders and containers.
Login now checks to see if the user who logged in is not the previous
user who logged in. If so, it will erase all of the objects to be
uploaded and all of the images taken with the internal camera.
Login now checks to see if the user who logged in last is the current
user. If so, does not changed much.

Note - Create an alert to inform the user, if the user logging in is a
different user from the one previously logged in, that the information
on the device will be erased.