AutoImageDelete & Wifi Only & Fixes for TaggingOverview
Implemented the auto delete feature — need to test. Implemented wifi
only feature — need to test and create an easy method to see if it’s
working.

Having issues with the overview of the images taken with the camera.
The blue outlines are not selecting the correct cells, but the correct
cells are actually being selected — Will fix issue later, need a break
from it. Lol.