Dashboard & Tagging
Dashboard now allows you to delete individual items. Dashboard now
tells you if there are any issues with permissions. Dashboard now will
update permissions when the app is not in the foreground and something
is changed.

Tagging page is now much more efficient when processing images already
taken with the application and sending them to the Dashboard